chore(folder-rename): wrong folder name
