+++ README.md
@@ -84,6 +84,7 @@ import { Toaster } from '@signozhq/ui';
 import { Toggle } from '@signozhq/ui';
 import { ToggleGroup } from '@signozhq/ui';
 import { Tooltip } from '@signozhq/ui';
+import { Typography } from '@signozhq/ui';
 ```
 
 ## Contributing
