+++ apps/docs/stories/intro.mdx
@@ -88,4 +88,5 @@ import { Toaster } from '@signozhq/ui';
 import { Toggle } from '@signozhq/ui';
 import { ToggleGroup } from '@signozhq/ui';
 import { Tooltip } from '@signozhq/ui';
+import { Typography } from '@signozhq/ui';
 ```
