+++ packages/ui/src/index.ts
@@ -21,10 +21,10 @@ export * from './resizable/index.js';
 export * from './select/index.js';
 export * from './sonner/index.js';
 export * from './switch/index.js';
-export * from './typograhy/index.js';
 export * from './table/index.js';
 export * from './tabs/index.js';
 export * from './text-ellipsis/index.js';
 export * from './toggle/index.js';
 export * from './toggle-group/index.js';
 export * from './tooltip/index.js';
+export * from './typography/index.js';
